Refactor: Rental 서비스 인터페이스 제거 및 구현체 직접 사용
- 향후 확장성 및 추상화의 필요성이 낮다고 판단되어 불필요한 서비스 인터페이스를 제거하고, 해당 구현체를 직접 사용하도록 변경했습니다.
