+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/rentalreservation/application/service/RentalQueryService.java
@@ -1,33 +1,61 @@
 package com.nextdoor.nextdoor.domain.rentalreservation.application.service;
 
 import com.nextdoor.nextdoor.domain.rentalreservation.application.dto.*;
+import com.nextdoor.nextdoor.domain.rentalreservation.application.port.RentalQueryPort;
+import com.nextdoor.nextdoor.domain.rentalreservation.domain.model.RentalReservation;
+import com.nextdoor.nextdoor.domain.rentalreservation.domain.exception.NoSuchRentalException;
+import com.nextdoor.nextdoor.domain.rentalreservation.domain.repository.RentalReservationRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-/**
- * 대여 조회 서비스
- * 책임: 대여 예약 정보 조회 (읽기 전용) 작업을 담당합니다.
- */
 @Service
-public interface RentalQueryService {
+@RequiredArgsConstructor
+public class RentalQueryService {
 
-    /**
-     * 대여 목록을 검색합니다.
-     */
-    Page<SearchRentalResult> searchRentals(SearchRentalCommand command);
+    private final RentalReservationRepository rentalReservationRepository;
+    private final RentalQueryPort rentalQueryPort;
 
-    /**
-     * ID로 대여 정보를 조회합니다.
-     */
-    SearchRentalResult getRentalById(Long rentalId);
+    public Page<SearchRentalResult> searchRentals(SearchRentalCommand command) {
+        return rentalQueryPort.searchRentals(command);
+    }
 
-    /**
-     * 관리 중인 대여 수를 조회합니다.
-     */
-    ManagedRentalCountResult countManagedRentals(Long ownerId);
-    
-    /**
-     * 대여 정보를 삭제합니다.
-     */
-    DeleteRentalResult deleteRental(DeleteRentalCommand command);
+    @Transactional(readOnly = true)
+    public SearchRentalResult getRentalById(Long rentalId) {
+        return rentalQueryPort.findRentalById(rentalId)
+                .orElseThrow(() -> new NoSuchRentalException("ID가 " + rentalId + "인 대여 정보가 존재하지 않습니다."));
+    }
+
+    public ManagedRentalCountResult countManagedRentals(Long ownerId) {
+        return rentalQueryPort.countManagedRentals(ownerId);
+    }
+
+    @Transactional
+    public DeleteRentalResult deleteRental(DeleteRentalCommand command) {
+        try {
+            RentalReservation rental = rentalReservationRepository.findById(command.getRentalId())
+                    .orElseThrow(() -> new NoSuchRentalException("ID가 " + command.getRentalId() + "인 대여 정보가 존재하지 않습니다."));
+
+            rentalReservationRepository.delete(rental);
+
+            return DeleteRentalResult.builder()
+                    .rentalId(command.getRentalId())
+                    .success(true)
+                    .message("대여 정보가 성공적으로 삭제되었습니다.")
+                    .build();
+        } catch (NoSuchRentalException e) {
+            return DeleteRentalResult.builder()
+                    .rentalId(command.getRentalId())
+                    .success(false)
+                    .message(e.getMessage())
+                    .build();
+        } catch (Exception e) {
+            return DeleteRentalResult.builder()
+                    .rentalId(command.getRentalId())
+                    .success(false)
+                    .message("대여 정보 삭제 중 오류가 발생했습니다: " + e.getMessage())
+                    .build();
+        }
+    }
 }
